# modules/application/app/src/test/resources/openapi/mobile-client-openapi-snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nyx Media Streamer API",
    "version": "1.0.0",
    "description": "Adaptive media server: transcoding, images, audio, eForms"
  },
  "paths": {
...
    "/api/v1/browse": {
      "get": {
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Invalid sort or path parameter"
          }
        },
        "description": "Browse a directory of media files",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "description": "Virtual path to browse (empty for root)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dir",
            "in": "query",
            "required": false,
            "description": "Alias for path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number (0-based)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Items per page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "description": "Sort order: name, date, size",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/stream.m3u8": {
      "get": {
        "responses": {
          "200": {
            "description": "HLS master playlist (m3u8)"
          },
          "202": {
            "description": ""
          },
          "400": {
            "description": "Invalid path or quality parameter"
          }
        },
        "description": "Open an HLS playback session from a media path and return the master playlist when ready",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "Virtual path to the media file",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quality",
            "in": "query",
            "required": false,
            "description": "Quality preset: high, low, medium, or passthrough",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  }
}